requirements.txt

openai>=1.35.13
python-dotenv>=1.0.1
